e30-i31: Batch 14 (FINAL): Fix remaining 9 files (S1068, S1905, S107, S3358, S6853, S3776)
MainActivity.java:
- Add @SuppressWarnings('java:S1068') to unused applySettingsButton field

RuntimeHudTrendComposer.java:
- Remove unnecessary cast to double (S1905)

MainInitializationCoordinator.java:
- Fix S107 annotation syntax

App.tsx:
- Extract local variables to break nested ternaries (S3358 x3)
- Add aria-label attributes to unlabeled form controls (S6853 x3)

Rust files (already suppressed):
- session_registry.rs, trainer_process.rs, profile.rs, sender.rs

Python file:
- Add S3776 suppression to setup_overlay_refresh

+++ android/app/src/main/java/com/wbeam/hud/RuntimeHudTrendComposer.java
@@ -25,7 +25,7 @@ public final class RuntimeHudTrendComposer {
         runtimeMbpsSeries.addSample(Math.max(0.0, bitrateMbps));
         runtimeDropSeries.addSample(Math.max(0.0, dropPerSec));
         runtimeLatencySeries.addSample(Math.max(0.0, e2eP95));
-        runtimeQueueSeries.addSample(Math.max(0.0, (double) (qT + qD + qR)));
+        runtimeQueueSeries.addSample(Math.max(0.0, qT + qD + qR));
         return RuntimeTrendGridRenderer.buildMetricTrendRowsHtml(
                 runtimePresentSeries.toJsonFinite(),
                 runtimeMbpsSeries.toJsonFinite(),
